# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Moustass API",
        "description": "API sécurisée pour l'application Moustass",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://www.moustass.com/api",
            "description": "Serveur local"
        }
    ],
    "paths": {
        "/admin/users": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Liste tous les utilisateurs",
                "operationId": "15e661c1bccc3c5cfdf1c1c80c4f1ae4",
                "responses": {
                    "200": {
                        "description": "Liste des utilisateurs"
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "403": {
                        "description": "Accès refusé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "Créer un nouvel utilisateur",
                "operationId": "729b0fbecee9fae2c50352ff1d92744d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password",
                                    "role"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "bob@moustass.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "MonMotDePasse1!"
                                    },
                                    "role": {
                                        "type": "string",
                                        "enum": [
                                            "ADMIN",
                                            "CLIENT"
                                        ],
                                        "example": "CLIENT"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Utilisateur créé avec succès"
                    },
                    "400": {
                        "description": "Données invalides"
                    },
                    "403": {
                        "description": "Accès refusé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/users/{id}": {
            "put": {
                "tags": [
                    "Admin"
                ],
                "summary": "Modifier un utilisateur",
                "operationId": "aff37f1fba67c923cb46f1779039e4a4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "role",
                                    "status"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "bob@moustass.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "NouveauMotDePasse1!"
                                    },
                                    "role": {
                                        "type": "string",
                                        "enum": [
                                            "ADMIN",
                                            "CLIENT"
                                        ]
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "active",
                                            "disabled"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Utilisateur modifié avec succès"
                    },
                    "404": {
                        "description": "Utilisateur introuvable"
                    },
                    "403": {
                        "description": "Accès refusé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Admin"
                ],
                "summary": "Supprimer un utilisateur",
                "operationId": "372c9e97299a21412d67f0623124cf2a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Utilisateur supprimé avec succès"
                    },
                    "403": {
                        "description": "Action non autorisée"
                    },
                    "404": {
                        "description": "Utilisateur introuvable"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/users/{id}/disable": {
            "patch": {
                "tags": [
                    "Admin"
                ],
                "summary": "Désactiver un utilisateur",
                "operationId": "a2bf82607de41b60811e9cc1768bb2d7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Utilisateur désactivé avec succès"
                    },
                    "404": {
                        "description": "Utilisateur introuvable"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/users/{id}/enable": {
            "patch": {
                "tags": [
                    "Admin"
                ],
                "summary": "Activer un utilisateur",
                "operationId": "9812bcb1655e334fa27dca1c6debc6f3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Utilisateur activé avec succès"
                    },
                    "404": {
                        "description": "Utilisateur introuvable"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}